# dashboard/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "DownloadCreateRequest": {
        "properties": {
          "model_name": {
            "title": "Model Name",
            "type": "string"
          }
        },
        "required": ["model_name"],
        "title": "DownloadCreateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelDownloadJob": {
        "description": "\u672c\u5730\u6a21\u578b\u4e0b\u8f7d\u4efb\u52a1\u5feb\u7167\u3002\n\n\u53ea\u8bb0\u5f55\u6a21\u578b\u540d\u3001\u8fdb\u5ea6\u548c\u9519\u8bef\u6458\u8981\uff1b\u6a21\u578b\u4e8c\u8fdb\u5236\u4ecd\u7531 Ollama \u81ea\u5df1\u7ba1\u7406\u3002",
        "properties": {
          "cancel_requested": {
            "default": false,
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "completed_bytes": {
            "default": 0,
            "title": "Completed Bytes",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_layer": {
            "default": "",
            "title": "Current Layer",
            "type": "string"
          },
          "digest": {
            "default": "",
            "title": "Digest",
            "type": "string"
          },
          "download_id": {
            "title": "Download Id",
            "type": "string"
          },
          "error_code": {
            "default": "",
            "title": "Error Code",
            "type": "string"
          },
          "error_message": {
            "default": "",
            "title": "Error Message",
            "type": "string"
          },
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "progress_ratio": {
            "default": 0.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress Ratio",
            "type": "number"
          },
          "provider": {
            "default": "OLLAMA",
            "title": "Provider",
            "type": "string"
          },
          "requested_by": {
            "default": "",
            "title": "Requested By",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/ModelDownloadStatus",
            "default": "CREATED"
          },
          "total_bytes": {
            "default": 0,
            "title": "Total Bytes",
            "type": "integer"
          }
        },
        "required": ["download_id", "model_name"],
        "title": "ModelDownloadJob",
        "type": "object"
      },
...
      "ModelProviderProfile": {
        "description": "\u975e\u654f\u611f\u6a21\u578b profile\u3002",
        "properties": {
          "active": {
            "default": false,
            "title": "Active",
            "type": "boolean"
          },
          "api_key": {
            "title": "Api Key",
            "type": "null"
          },
          "base_url": {
            "default": "",
            "title": "Base Url",
            "type": "string"
          },
          "chat_completions_path": {
            "default": "/v1/chat/completions",
            "title": "Chat Completions Path",
            "type": "string"
          },
          "config_version": {
            "default": 1,
            "title": "Config Version",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "endpoint_hash": {
            "default": "",
            "title": "Endpoint Hash",
            "type": "string"
          },
          "json_mode": {
            "default": true,
            "title": "Json Mode",
            "type": "boolean"
          },
          "max_retries": {
            "default": 2,
            "title": "Max Retries",
            "type": "integer"
          },
          "max_tokens": {
            "default": 4096,
            "title": "Max Tokens",
            "type": "integer"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "profile_id": {
            "title": "Profile Id",
            "type": "string"
          },
          "provider_kind": {
            "$ref": "#/components/schemas/PlannerProviderKind"
          },
          "secret_present": {
            "default": false,
            "title": "Secret Present",
            "type": "boolean"
          },
          "secret_store_kind": {
            "$ref": "#/components/schemas/SecretStoreKind",
            "default": "SESSION_ONLY"
          },
          "secret_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret Updated At"
          },
          "temperature": {
            "default": 0.0,
            "title": "Temperature",
            "type": "number"
          },
          "timeout_seconds": {
            "default": 30.0,
            "title": "Timeout Seconds",
            "type": "number"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "profile_id",
          "display_name",
          "provider_kind",
          "model_name"
        ],
        "title": "ModelProviderProfile",
        "type": "object"
      },
...
      "PlannerDryRunRequest": {
        "properties": {
          "control_mode": {
            "title": "Control Mode",
            "type": "string"
          },
          "sample_scene": {
            "title": "Sample Scene",
            "type": "string"
          },
          "user_instruction": {
            "title": "User Instruction",
            "type": "string"
          }
        },
        "required": ["user_instruction", "sample_scene", "control_mode"],
        "title": "PlannerDryRunRequest",
        "type": "object"
      },
      "PlannerProviderKind": {
        "description": "\u652f\u6301\u7684 Planner provider \u7c7b\u578b\u3002",
        "enum": ["MOCK", "RULE_BASED", "OPENAI_COMPATIBLE", "OLLAMA"],
        "title": "PlannerProviderKind",
        "type": "string"
      },
      "PlannerRuntimeStatus": {
        "description": "\u5f53\u524d active planner \u72b6\u6001\u3002",
        "properties": {
          "active_model": {
            "default": "mock",
            "title": "Active Model",
            "type": "string"
          },
          "active_profile_id": {
            "default": "",
            "title": "Active Profile Id",
            "type": "string"
          },
          "active_provider": {
            "$ref": "#/components/schemas/PlannerProviderKind",
            "default": "MOCK"
          },
          "circuit_breaker": {
            "default": "CLOSED",
            "title": "Circuit Breaker",
            "type": "string"
          },
          "config_version": {
            "default": 0,
            "title": "Config Version",
            "type": "integer"
          },
          "endpoint_hash": {
            "default": "",
            "title": "Endpoint Hash",
            "type": "string"
          },
          "health": {
            "default": "READY",
            "title": "Health",
            "type": "string"
          },
          "last_test": {
            "additionalProperties": true,
            "title": "Last Test",
            "type": "object"
          }
        },
        "title": "PlannerRuntimeStatus",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "BIG-small Cloud Planning API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/model-control/ollama/downloads": {
      "get": {
        "operationId": "list_ollama_downloads_api_v1_model_control_ollama_downloads_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelDownloadJob"
                  },
                  "title": "Response List Ollama Downloads Api V1 Model Control Ollama Downloads Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Ollama Downloads",
        "tags": ["model-control"]
      },
      "post": {
        "operationId": "create_ollama_download_api_v1_model_control_ollama_downloads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDownloadJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Ollama Download",
        "tags": ["model-control"]
      }
    },
    "/api/v1/model-control/ollama/downloads/{download_id}": {
      "get": {
        "operationId": "get_ollama_download_api_v1_model_control_ollama_downloads__download_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "download_id",
            "required": true,
            "schema": {
              "title": "Download Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDownloadJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ollama Download",
        "tags": ["model-control"]
      }
    },
    "/api/v1/model-control/ollama/downloads/{download_id}/cancel": {
      "post": {
        "operationId": "cancel_ollama_download_api_v1_model_control_ollama_downloads__download_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "download_id",
            "required": true,
            "schema": {
              "title": "Download Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDownloadJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Ollama Download",
        "tags": ["model-control"]
      }
    },
...
    "/api/v1/model-control/ollama/models/{model_name}": {
      "delete": {
        "operationId": "delete_ollama_model_api_v1_model_control_ollama_models__model_name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "model_name",
            "required": true,
            "schema": {
              "title": "Model Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Ollama Model Api V1 Model Control Ollama Models  Model Name  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Ollama Model",
        "tags": ["model-control"]
      },
      "get": {
        "operationId": "ollama_model_detail_api_v1_model_control_ollama_models__model_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "model_name",
            "required": true,
            "schema": {
              "title": "Model Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ollama Model Detail Api V1 Model Control Ollama Models  Model Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ollama Model Detail",
        "tags": ["model-control"]
      }
    },
    "/api/v1/model-control/ollama/models/{model_name}/activate": {
      "post": {
        "operationId": "activate_ollama_model_api_v1_model_control_ollama_models__model_name__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "model_name",
            "required": true,
            "schema": {
              "title": "Model Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannerRuntimeStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate Ollama Model",
        "tags": ["model-control"]
      }
    },
...
    "/api/v1/model-control/planner/dry-run": {
      "post": {
        "operationId": "planner_dry_run_api_v1_model_control_planner_dry_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlannerDryRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Planner Dry Run Api V1 Model Control Planner Dry Run Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Planner Dry Run",
        "tags": ["model-control"]
      }
    },
    "/api/v1/model-control/profiles": {
      "get": {
        "operationId": "list_profiles_api_v1_model_control_profiles_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelProviderProfile"
                  },
                  "title": "Response List Profiles Api V1 Model Control Profiles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Profiles",
        "tags": ["model-control"]
      },
      "post": {
        "operationId": "create_profile_api_v1_model_control_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProviderProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Profile",
        "tags": ["model-control"]
      }
    },
    "/api/v1/model-control/profiles/{profile_id}": {
      "delete": {
        "operationId": "delete_profile_api_v1_model_control_profiles__profile_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Profile",
        "tags": ["model-control"]
      },
      "patch": {
        "operationId": "update_profile_api_v1_model_control_profiles__profile_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProviderProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Profile",
        "tags": ["model-control"]
      }
    },
    "/api/v1/model-control/profiles/{profile_id}/activate": {
      "post": {
        "operationId": "activate_profile_api_v1_model_control_profiles__profile_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlannerRuntimeStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate Profile",
        "tags": ["model-control"]
      }
    },
    "/api/v1/model-control/profiles/{profile_id}/test": {
      "post": {
        "operationId": "test_profile_api_v1_model_control_profiles__profile_id__test_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Test Profile Api V1 Model Control Profiles  Profile Id  Test Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Profile",
        "tags": ["model-control"]
      }
    },
...
  }
}